Django/cloudstorage-demo/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
django==5.2.7
django-jazzmin==3.0.1
django-ninja==1.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
#psycopg2==2.9.10
pydantic==2.12.0
pydantic-core==2.41.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
#starlette==0.48.0
starlette
typing-extensions==4.15.0
typing-inspection==0.4.2
#uwsgi==2.0.30
